Apply pyupgrade fixes for Python 3.9+ syntax (#8150)
### Description

Included is a commit here applying pyupgrade changes for Python 3.8+
syntax. This should have been included in
https://github.com/Project-MONAI/MONAI/commit/2b19d9c2a28e5aa432d54f61728f24047e60313e
when Python 3.7 support was dropped.

Also included is a commit here apply pyupgrade changes for Python 3.9+
syntax. This should have been included in
https://github.com/Project-MONAI/MONAI/commit/7b80e12b6eddb7d929c92c1a3d6557c8724cc89f
when Python 3.8 support was dropped.

I've also run the pre-commit autoupdate command to use the latest
versions of the various pre-commit hook repos. It appears that
pre-commit.ci bot has not been doing this on the quarterly schedule as
expected? It appears the last time it submitted the PR to update the
pre-commit hook repos was back in
https://github.com/Project-MONAI/MONAI/pull/6286 from April 2023.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).

---------

Signed-off-by: James Butler <james.butler@revvity.com>